==> analysis/scripts/1_data_preparation.ipynb <==
import pandas as pd
import numpy as np
import json

**Preparing the dataset - processing exported CSV files into accessible format.**
- input: raw exports from UXtweak study
- output: formatted respondent and result data

# Respontents and questionnaires formatting

## Respondents formatting

**Load raw files.**

respondents_2s = pd.read_csv("../data/raw_data/5ST case study - 2 seconds - Respondents.csv")
respondents_5s = pd.read_csv("../data/raw_data/5ST case study - 5 seconds - Respondents.csv")
respondents_10s = pd.read_csv("../data/raw_data/5ST case study - 10 seconds - Respondents.csv")
prestudy_2s = pd.read_csv("../data/raw_data/5ST case study - 2 seconds - Pre-study questionnaire.csv")
prestudy_5s = pd.read_csv("../data/raw_data/5ST case study - 5 seconds - Pre-study questionnaire.csv")
prestudy_10s = pd.read_csv("../data/raw_data/5ST case study - 10 seconds - Pre-study questionnaire.csv")
poststudy_2s = pd.read_csv("../data/raw_data/5ST case study - 2 seconds - Post-study questionnaire.csv")
poststudy_5s = pd.read_csv("../data/raw_data/5ST case study - 5 seconds - Post-study questionnaire.csv")
poststudy_10s = pd.read_csv("../data/raw_data/5ST case study - 10 seconds - Post-study questionnaire.csv")

respondents_5s.head()

prestudy_5s.head()

poststudy_5s.head()

respondents_output_2s = respondents_2s.merge(prestudy_2s, how='outer', left_on='order', right_on='respondent').merge(poststudy_2s, how='outer', left_on='order', right_on='respondent')
respondents_output_5s = respondents_5s.merge(prestudy_5s, how='outer', left_on='order', right_on='respondent').merge(poststudy_5s, how='outer', left_on='order', right_on='respondent')
respondents_output_10s = respondents_10s.merge(prestudy_10s, how='outer', left_on='order', right_on='respondent').merge(poststudy_10s, how='outer', left_on='order', right_on='respondent')

respondents_output_5s.columns

**Merging files.**

respondents_output_2s['interval'] = '2s'
respondents_output_5s['interval'] = '5s'
respondents_output_10s['interval'] = '10s'

respondents = pd.concat([respondents_output_2s, respondents_output_5s, respondents_output_10s], axis=0).reset_index(drop=True)
respondents.head()

**Formatting cognitive abilities tasks results.**

for index, row in respondents.iterrows():
   if(type(row['VMT attempts']) == str):
        field = json.loads(row['VMT attempts'])
        output = []
        for f in field:
            output.append([])
            for x in f['clicks']:
                output[len(output)-1].append({
                    'type': x['type'],
                    'time': x['timestamp']
                })
        respondents.loc[index, 'VMT attempts'] = json.dumps(output)
   if(type(row['PST levels']) == str):     
        field2 = json.loads(row['PST levels'])
        output2 = [stamp['canClick'] for stamp in field2]
        respondents.loc[index, 'PST levels'] = json.dumps(output2)

**Renaming columns.**

columns={
    'order': 'respondent',
    'included in analysis': 'included',
    'started at': 'started_at',
    'time taken': 'time_taken',
    'questions answered': 'questions_answered',
    'tasks completed': 'tasks_completed',
    'tasks skipped': 'tasks_skipped',
    'Q1: What devices are you using right now?': 'input_device',
    'Q2: What age group do you belong to?': 'age',
    'Q3: What gender do you identify as?': 'gender',
    'Q4: What is the highest education level you have completed?': 'education',
    'Q5: What is your personal income per year, after tax?': 'income',
    'Q6: How often do you browse the web?': 'web_browsing',
    'Q7: What do you think is the level of your skill as a user on the web?': 'web_skill',
    'Q8: You may love pizza or ice cream. But when we ask you what you would order in the restaurant, you need to choose a salad.': 'att_check_1',
    'Q1: During the last 12 months, have you participated in any studies similar to this one where you were instructed to complete tasks and answer questions afterwards? (original)': 'similar_studies',
    'Q2: Is there anything else you’d like to tell us? (original)' : 'note',
    'Q3: How many months are there in a year? Even though the correct answer is twelve, make sure to choose the option thirty.' : 'att_check_2',
    'VMT achieved level': 'vmt_levels',
    'VMT attempts': 'vmt_clicks',
    'PST levels passed': 'pst_levels',
    'PST levels correct': 'pst_levels_correct',
    'PST levels': 'pst_clicks'
}

drop=[
    'respondent_x',
    'respondent_y',
    'Q1: During the last 12 months, have you participated in any studies similar to this one where you were instructed to complete tasks and answer questions afterwards? (grouped)',
    'Q2: Is there anything else you’d like to tell us? (grouped)'
]

respondents = respondents.rename(columns=columns)

respondents.columns

respondents = respondents.drop(drop, axis=1)

**Column types.**

respondents['started_at'] = respondents['started_at'].astype('datetime64')
respondents['time_taken'] = pd.to_timedelta(respondents.time_taken).astype('timedelta64[m]')

**Export.**

respondents.to_csv("../data/respondents_to_calculate.csv")

## Raw results formatting

**Load data.**

labels = ['company', 'purpose', 'detail', 'description', 'PQ1', 'ATT1', 'PQ2', 'HQ1', 'PQ3', 'HQ2', 'HQ3', 'ATT2', 'PQ4', 'HQ4', 'familiarity']

results_2s = pd.read_csv("../data/raw_data/5ST case study - 2 seconds - Raw results.csv")
results_5s = pd.read_csv("../data/raw_data/5ST case study - 5 seconds - Raw results.csv")
results_10s = pd.read_csv("../data/raw_data/5ST case study - 10 seconds - Raw results.csv")

results_5s.columns

results_5s.head()

**Remove unecessary data.**

results_2s = results_2s.drop(results_2s.columns[[2,3,4,6,7,8,9,10]], axis=1)
results_5s = results_5s.drop(results_5s.columns[[2,3,4,6,7,8,9,10]], axis=1)
results_10s = results_10s.drop(results_10s.columns[[2,3,4,6,7,8,9,10]], axis=1)

results_2s = results_2s.drop(results_2s[(results_2s['task number'] == 1)].index)
results_5s = results_5s.drop(results_5s[(results_5s['task number'] == 1)].index)
results_10s = results_10s.drop(results_10s[(results_10s['task number'] == 1)].index)

**New result dataframes with formatted data.**

results_2s_final = pd.DataFrame()
results_2s_final['respondent'] = results_2s['respondent number']
results_2s_final['task'] = results_2s['task number']
results_2s_final['interval'] = '2s'
results_2s_final['time'] = results_2s['time taken']

results_5s_final = pd.DataFrame()
results_5s_final['respondent'] = results_5s['respondent number']
results_5s_final['task'] = results_5s['task number']
results_5s_final['interval'] = '5s'
results_5s_final['time'] = results_5s['time taken']

results_10s_final = pd.DataFrame()
results_10s_final['respondent'] = results_10s['respondent number']
results_10s_final['task'] = results_10s['task number']
results_10s_final['interval'] = '10s'
results_10s_final['time'] = results_10s['time taken']

for i in range(15):
    res_2s = pd.Series()
    res_5s = pd.Series()
    res_10s = pd.Series()
    
    for j in range(6):
        if (res_2s.size == 0):
            res_2s = results_2s.iloc[:, 3 + j*15 + i].fillna('').astype(str)
        else:
            res_2s += results_2s.iloc[:, 3 + j*15 + i].fillna('').astype(str)
        
        if (res_5s.size == 0):
            res_5s = results_5s.iloc[:, 3 + j*15 + i].fillna('').astype(str)
        else:
            res_5s += results_5s.iloc[:, 3 + j*15 + i].fillna('').astype(str)
        
        if (res_10s.size == 0):
            res_10s = results_10s.iloc[:, 3 + j*15 + i].fillna('').astype(str)
        else:
            res_10s += results_10s.iloc[:, 3 + j*15 + i].fillna('').astype(str)
        
    results_2s_final[labels[i]] = res_2s
    results_5s_final[labels[i]] = res_5s
    results_10s_final[labels[i]] = res_10s

results_2s_final = results_2s_final.sort_values(by=['respondent', 'task']).reset_index(drop=True)
results_5s_final = results_5s_final.sort_values(by=['respondent', 'task']).reset_index(drop=True)
results_10s_final = results_10s_final.sort_values(by=['respondent', 'task']).reset_index(drop=True)

results = pd.concat([results_2s_final, results_5s_final, results_10s_final], axis=0).reset_index(drop=True)
results.head()

results['task'] = results['task'] - 1

results['familiarity'] = results['familiarity'].replace(to_replace='No', value=False)
results['familiarity'] = results['familiarity'].replace(to_replace='Yes', value=True)

results.head()

results.columns

**Export.**

results.to_csv("../data/results_unlabeled.csv")
